backend/backend_requirements.txt
# Backend Requirements - requirements.txt

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# Background Tasks
celery==5.3.4
redis==5.0.1

# HTTP Client
httpx==0.25.2
requests==2.31.0

# ML & Analytics
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Monitoring & Logging
structlog==23.2.0
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Environment
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Validation
email-validator==2.1.0

# Rate Limiting
slowapi==0.1.9

# Caching
aioredis==2.0.1